fixup! refactor: move EventName to the event module
